requirements.txt
google-adk>=1.9.0
google-cloud-bigquery>=3.35.1
google-genai>=1.28.0
google-cloud-storage>=2.10.0 